+++ ChangeLog.md
@@ -5,8 +5,6 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 # TO-DO List
 
 - Add command `/log copy <player>` or `/logCopy <player>`, to copy logger configurations from other players.
-- Display the actual option when setting logger with default options.
-    - `Subscribed to <logger> \(<option>\)`
 - Bump to release versions after merging pathfinding-logger?
 - Improve SpawnReporter: to report passive mob cap every tick.
 - Add another mode of mob-cap logger.
@@ -15,6 +13,7 @@ Carpet 1.12 with [RNY](https://github.com/Rainyaphthyl)'s Addition
 
 - Fix a display bug about symbol `%` of command `/spawn` on the server console.
 - Add `pathFinding` logger, displaying entity AI paths with particle lines.
+- Display the actual default option when setting logger with null options, i.e. `Subscribed to <logger> \(<option>\)`.
 
 # Code Details